+++ examples/with-vpc/main.tf
@@ -30,7 +30,7 @@ module "lambda_function_in_vpc" {
 
 module "vpc" {
   source  = "terraform-aws-modules/vpc/aws"
-  version = "~> 5.0"
+  version = "~> 6.0"
 
   name = random_pet.this.id
   cidr = "10.10.0.0/16"